Move blacklist client code into module
