+++ modules/ept-blacklist/routes/all.js
@@ -14,7 +14,7 @@
   */
 module.exports = {
   method: 'GET',
-  path: '/api/admin/settings/blacklist',
+  path: '/api/admin/blacklist',
   config: {
     auth: { strategy: 'jwt' },
     pre: [ { method: 'auth.blacklist.all(server, auth)' } ]
