+++ modules/ept-frontend/client/resources/admin/settings.js
@@ -4,27 +4,6 @@
 module.exports = ['$resource',
   function($resource) {
     return $resource('/api/admin/settings/', {}, {
-      getBlacklist: {
-        method: 'GET',
-        url: '/api/admin/settings/blacklist',
-        isArray: true
-      },
-      addToBlacklist: {
-        method: 'POST',
-        url: '/api/admin/settings/blacklist',
-        isArray: true
-      },
-      updateBlacklist: {
-        method: 'PUT',
-        url: '/api/admin/settings/blacklist',
-        isArray: true
-      },
-      deleteFromBlacklist: {
-        method: 'DELETE',
-        url: '/api/admin/settings/blacklist/:id',
-        params: { id: '@id' },
-        isArray: true
-      },
       getTheme: {
         method: 'GET',
         url: '/api/admin/settings/theme'
